updated with genesis file, added new bootnodes
Also had to change the fork check, as we were putting previous and current as the genesis fork, which didn't compare favourably with actual genesis.

I had to add a toString helper to see clearly what was different, so i left that as the fork data in the error message if we get it - it'll let us see all 3 fields and it was actually the 1 field i hadn't printed that was different.

Signed-off-by: Paul Harris <paul.harris@consensys.net>