Reverts the state selection changes
Refactored the generation of attestations. This will more memory, but the payoff is pretty large.

ran my test script to generate 50k duties for 'next epoch' on our archive holesky node, script took 17 seconds (not same as my machine, my machine pre-change was about 12 seconds).

Post change on my laptop, queried the same epoch as on nightly (so that i could verify cksum of output on a real network, because the change was a huge amount faster)

My machine computed in 0.24 seconds. Better.

In terms of the JMH, went from roughly 1op/s to roughly 1250 op/s.

I considered doing the old method for smaller requests but this is considerably faster even for a few hundred keys, so will use the new method for all requests.

Signed-off-by: Paul Harris <paul.harris@consensys.net>