teku-5930 Make it the default to not sync from Genesis
One common issue when starting a new node is start synchronizing from genesis, instead of using checkpoint sync. Not only this takes longer to sync, but is riskier (subject to long-range attacks).

By default, Teku won't allow users to sync from an epoch that is too old (out of the weak subjectivity period). This includes trying to sync from genesis on mainnet.

For most users, the best option to start a new node is to use [`--checkpoint-sync-url`](https://docs.teku.consensys.net/development/get-started/checkpoint-start) pointing to a checkpoint sync endpoint in sync with the network. This is going to start their node from the latest finalized checkpoint.

This PR added a CLI option `--ignore-weak-subjectivity-period-enabled` to allow users to opt-out of this behaviour.