Implemented get_proposer_head for late block reorg
Implemented get_proposer_head function, and tested results via the reftests.

It's going to be more efficient to get the justified state once, so have refactored isHeadWeak and isParentStrong.

The ordering of checks was cheapest -> most expensive, we can avoid fetching state if any of the lighter tests fail.

partially addresses #6595

Signed-off-by: Paul Harris <paul.harris@consensys.net>