Implemented should_override_fork_choice for late block reorg
Implemented should_override_fork_choice function, and tested results via the reftests.

There's a level of duplication in the 'should_override_fork_choice' and 'get_proposer_head', but I've stayed true to the spec, while fixing ordering a little to do cheaper tests first.

Ultimately, the function will exit fast once it finds a mis-match, and if it gets to the end then things pass.

Also had to adjust timeliness to use store time, and I made the Store implement the TimeProvider to fix an issue with using a system time provider when store time was needed.

Finally, while I did successfully run the reftests, I have to leave the reftests disabled until we get a release that doesnt contain backwards ticks. It's possible to remove the ticks to run the tests, but they won't run out of the box.

partially addresses #6595

Signed-off-by: Paul Harris <paul.harris@consensys.net>