Removed synchronized section in localSlashingProtector
The idea with this change is to reduce contention on the synchronized functions, which means we need to introduce more locking, but this will allow better resource usage.  Files etc are still written synchronously, but we'll now be able to check more than one signing at a time, as long as they're distinct validators.

Roughly related to #7663

Signed-off-by: Paul Harris <paul.harris@consensys.net>