Read entries of an ERA file
Currently can't unpack the framed snappy data, but this is a logical starting point which implements the reading of indexes and iterates the objects.

Included the python and a reference to the markdown that defines the file format.

This is not plugged into anything, currently just a work in progress with a static main fn. but has enough that we should save this as a starting point for the next iteration.

Signed-off-by: Paul Harris <paul.harris@consensys.net>