Skip fork choice update if it has been processed
Basically when preparing for attestations, we re-run FcU, and while we try to ensure this only happens once, we should be able to not run it at all now that FcU is run during block processing.

We should double check this on a testnet just to make sure the assertion that we can skip this extra FcU is correct. If it's an incorrect assertion we'd expect to see a significant decrease in finding the correct head in attestations.

fixes #7814

Signed-off-by: Paul Harris <paul.harris@consensys.net>