Added header to builder call
When calling `/eth/v1/builder/blinded_blocks` we had missed the Eth-Consensus-Version header.

I changed the OkHttpRestClient from using HttpUrl as there was a warning on the interface we were using.

We now add the header. I manually lower-cased the milestone string, as the version method we use in schema isn't really what we want in this context, but at some point we probably want to build this logic into SpecMilestone rather than relying on the Schema Version object.

fixes #8061.

Signed-off-by: Paul Harris <paul.harris@consensys.net>