Allow voluntary exit command to write to file rather than publish
The advantage of this is that the message can avoid epoch validation, and also it can separate the submission of the exit from creation, so if something fails, it can have further investigation.

The format saved to json will allow it to be posted directly to the voluntary exits beacon-api POST endpoint.

Assists with #8158 but only as a potential workaround, allowing us to avoid setting up the keymanager-api to generate an exit.

Signed-off-by: Paul Harris <paul.harris@consensys.net>