reverted validatorIndexCache
Reverted the changes to validatorIndexCache, we'll have to take a look again for full electra support.

Although the intent was to only index finalized validators, the result was to constantly be doing StateTransitions somehow to get indices, and the outcome was a serious slow down in the important state transitions that happen constantly, so as a starting point, this will put us back to releasable code that will be ok for Deneb.

Before reverting, it seemed to process 1 state transition every 20 seconds or so, and even just rebuilding a handful of states was taking 10 minutes or longer, which is not really going to work in production.

We'll also need to build out testing for this once we re-implement. Archive storage is already slow due to epoch processing, we can't afford to take this kind of increased performance hit.

I'm guessing the tracker may still be part of an eventual solution, haven't removed at this point but i've cleaned it out functionally... i probably need to clean that before any merge...

`BlockProcessorElectra` was changed to not use the cache to find validator index, as it's a simple operation, and it was causing problems.

Signed-off-by: Paul Harris <paul.harris@consensys.net>