Added deneb remote signer AT
Had to leave this test disabled, because it needs a change not currently in web3signer develop. Once it's in develop, we can start adding it in.

Also changed so that trusted-setup always defaults to mainnet if it's not set, which will remove the need in the first place for the web3signer change, but to be able to test this is the current easiest path i could see.

Signed-off-by: Paul Harris <paul.harris@consensys.net>