Expanded VC API tests
fixes #8756

Added test cases for VC block requests, and added a validation to the Attestation request test. These were the 2 VC cases where we now require that header to be specified or we will fail.

There was another case in DSL for AT, but we'll see those quickly if we use it.

Signed-off-by: Paul Harris <paul.harris@consensys.net>