Save earliest slot block in a DB variable (#8722)
Adding a variable in the database to store the earliest available block slot which is used when peers request blocks by range. This is a query that can vary in some case and I've seen it causing disconnection in some cases due to multiple requests getting timed out. Changing it to be stored in a variable brings the query to 1 - 2 ms response time. It does add a bit of logic to handle the sync of historical batches and prune which are currently the workflow in the code base that affect that column family.

Signed-off-by: Gabriel Fukushima <gabrielfukushima@gmail.com>

---------

Signed-off-by: Gabriel Fukushima <gabrielfukushima@gmail.com>