Added block production restrictions for electra
voluntary exits should not be included in a block if
 - they have a withdrawal pending (thats invalid)
 - they have a pending deposit (looking to make that invalid)

The pending withdrawal was pointed out in devnet5, so this resolves that.

The pending deposit is less cheap than i'd like ,but im fairly sure it'd potentially lead to some bad state data that we don't really want (eg. potentially depositing onto a withdrawn validator, so a withdrawn validator ending up with non 0 balance).

Signed-off-by: Paul Harris <paul.harris@consensys.net>