cleanup stale todo's
Removed the SlashingProtectedValidatorSource todo as its had several PRs from bots, and it's actually grossly over-simplifying.

Re-worded to remove TODO in EraFile, as its just a POC, and we won't be accepting any tiny changes there, the note was for when its rewritten.

Removed the TODO from SszCollectionSchema but left the content, as it does seem valid.

remove stale reference to a PR #3356
Fixes #7532

Signed-off-by: Paul Harris <paul.harris@consensys.net>