fix SerializableArrayTypeDefinition minItems and add maxItems
When I was looking at 9022, realised that minItems was hardcoded at a point and with a very specific use-case, so i've removed that in favor of a correct implementation.

Also implemented maxItems, and validations around min and max being at least 0 and in the right order of they'll generate a configuration exception.

I think it's fair to say this is basically only documentation currently given to nothing breaking other than our types tests.

I let this run and did some queries on a testnet and it was ok. will see if any AT's trip.

Partially addresses #9022

Signed-off-by: Paul Harris <paul.harris@consensys.net>