Ignore aggregate if falling behind
Currently if we're falling behind and receive an aggregate we attempt to validate it but our state is too old in some circumstances.

We can't validate this aggregate, but currently we're rejecting it, which is worse than ignoring it for gossip rules, so we're better off just ignoring it.

fixes #9187

Signed-off-by: Paul Harris <paul.harris@consensys.net>