Increased attestation queue limits
Also quashed RejectedExecution on attestation processing, as it was causing excessive disk load because of the volume, and it really does increase load a lot when it occurs.

The queue size should reduce the chance of hitting that, but the log change should assist if we do hit it. The point of rejectedExecution was not to actually increase load on the system.

fixes #9214

Signed-off-by: Paul Harris <paul.harris@consensys.net>