Added context to the sync committee duties error
The most plausible way for this error to have triggered is the node is not up to date with a chain, but we don't have the required context in the message to see what's going on.

It's good that the message has denied giving bad information, but almost equally important is the required context to understand why the request wasn't able to be fulfilled.

fixes #9348

Signed-off-by: Paul Harris <paul.harris@consensys.net>