check for next fork epoch, and error if we're beyond current fork.
The basic scenario is you'd be calling a miscHelper at an earlier fork than the one you need, so this check ensures we'd see an exception if we do that.

if you're following hte normal pattern of
```
spec.atEpoch(epoch).miscHelpers....
```

then you should get the right fork version and it shouldn't be a big issue, but if you're calling genesisSpec helpers, then your result would be misleading.