updated
fixed the location so it can access what's needed.

Figured out parameters for jmh, which improves how it all runs and reports.

We end up with 2 test suites, one that need precompute, and one that doesnt need precompute because it makes no difference. Should be easy to move functions if they change, as its all the same data underneath.