improved flow of POST voluntary exit
Improved messaging when exits are submitted that aren't valid, as we can pre-validate before we even attempt to access the cache

POST will now respond successful if it's ignoring due to the validator being already in the operation pool.

Added an info message to log when the operation has been seen.

fixes #9784