fixed get-variables command
Added in entries for missing variables

Changed to use builder for a few of the maps for consistency.

one of the columns was too large to be displayed nicely so changed the width of the variables display.

If a variable is missing, it will output as:
```
LATEST_CANONICAL_BLOCK_ROOT      : <NOT EXPORTED>
```

Ideally this would prompt a developer to add it to the variables mapping.

partially addresses #9845