Updated computeSubnetForAttestation
During non finality it's possible to have long periods without slots, and in these cases we would need to advance the state slot to be able to compute the attestation subnet.

Ordinarily this code will not be in range to be used, but we will need to start computing epochs to ensure we're in range.

This was found by diamond tests scenario nfin-001