Cleaned up recovery map as tasks complete
Added a timed task in case we miss items going through completed, but generally speaking the completed cleanup should finish things off nicely.

Also addressed the issue of storing the block in the recoveryEntry when we only need the slot and root.

I do think there's more to do here but focused on the stated problem.

fixes #9875