updated log levels in execution stub.
The goal was to have a less verbose output by default while still having a good level of information to see what's going on.

With everything at info it was so noisy on screen that it was hard to track what was going on.

Arguably this could be further tweaked to have some of the less common things at trace level, and then we can maybe still see when we want 'more' but not 'everything'.