increase download timeouts for sidecar retriever (#10190)
* increase download timeouts for sidecar retriever

This will help with recovery process.

Also changed logging a little for DasSampler to not just ignore timeouts, but currently just logging it explicitly as a warning.

Also removing cancelled downloads from the pending requests map.

The original recovery process had a 5-minute timeout for the downloads portion, we're up to 240s on download timeout after this change by default (from 40s). Having that increase seems to be related to longer range queries and having tons of column downloads in flight.

partially addresses #10185

* adjust recovery timeout for the acceptance test